Refactor #3 created PhotoSphereFeatures directory
