+++ src/buttons/EditNavMap.tsx
@@ -13,7 +13,7 @@ import {
 import { Photosphere, VFE } from "../DataStructures";
 import PhotosphereSelector, {
   PhotosphereSelectorProps,
-} from "../PhotosphereSelector";
+} from "../PhotoSphereFeatures/PhotosphereSelector";
 import { alertMUI } from "../UI/StyledDialogWrapper";
 
 interface EditNavMapProps {
